requirements.txt
requests
schedule
python-telegram-bot==21.0.1
